+++ modules/snapi-client/src/test/scala/raw/compiler/rql2/tests/output/JsonOutputTest.scala
@@ -13,7 +13,7 @@
 package raw.compiler.rql2.tests.output
 
 import raw.compiler.rql2.tests.CompilerTestContext
-import raw.utils._
+import raw.client.utils._
 
 import java.nio.file.{Files, Path}
